Retry transaction; custom config file and db file location; custom file icons for linux;

+++ BackBoxGui/src/it/backbox/gui/utility/FileUtility.java
@@ -2,12 +2,15 @@ package it.backbox.gui.utility;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URL;
 import java.util.HashMap;
+import java.util.Locale;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import javax.swing.Icon;
+import javax.swing.ImageIcon;
 import javax.swing.UIManager;
 import javax.swing.filechooser.FileSystemView;
 
@@ -16,32 +19,52 @@ import it.backbox.gui.BackBoxGui;
 public class FileUtility {
 	
 	private static final String FOLDER_KEY = "folder";
-
+	private static final String OS = System.getProperty("os.name").toLowerCase();
+	private static final boolean WINDOWS = OS.contains("windows");
 	private static final Logger _log = Logger.getLogger(BackBoxGui.class.getCanonicalName());
 	
 	private static Map<String, Icon> cacheIcon = new HashMap<>();
 	private static Map<String, String> cacheType = new HashMap<>();
 	private static FileSystemView fsv = FileSystemView.getFileSystemView();
 	private static File temp = new File(System.getProperty("java.io.tmpdir"));
+	private static Icon defaultIcon;
 	
 	public static Icon getIcon(String fileExt) {
+		fileExt = fileExt.toLowerCase(Locale.ROOT);
 		if (cacheIcon.containsKey(fileExt))
 			return cacheIcon.get(fileExt);
-		
-		File f = new File(temp, "icon." + fileExt);
-		try {
-			f.createNewFile();
-			f.deleteOnExit();
-			if (f.exists()) {
-				Icon i = fsv.getSystemIcon(f);
+		if (WINDOWS) {
+			// IT WORKS ONLY ON WINDOWS (Linux uses magic numbers to get file type)
+			File f = new File(temp, "icon." + fileExt);
+			try {
+				f.createNewFile();
+				f.deleteOnExit();
+				if (f.exists()) {
+					Icon i = fsv.getSystemIcon(f);
+					cacheIcon.put(fileExt, i);
+					return i;
+				}
+			} catch (IOException e) {
+				_log.log(Level.WARNING, "Error creating temp file for icon", e);
+			}
+		} else {
+			ClassLoader cl = Thread.currentThread().getContextClassLoader();
+			URL iconUrl = cl.getResource("images/fileicon/" + fileExt + ".png");
+			if (iconUrl != null) {
+				Icon i = new ImageIcon(iconUrl);
 				cacheIcon.put(fileExt, i);
 				return i;
 			}
-		} catch (IOException e) {
-			_log.log(Level.WARNING, "Error creating temp file for icon", e);
+			iconUrl = cl.getResource("images/fileicon/_blank.png");
+			if (iconUrl != null) {
+				Icon i = new ImageIcon(iconUrl);
+				cacheIcon.put(fileExt, i);
+				return i;
+			}
 		}
-		
-		return UIManager.getIcon("FileView.fileIcon");
+		if (defaultIcon == null)
+			defaultIcon = UIManager.getIcon("FileView.fileIcon");
+		return defaultIcon;
 	}
 
 	public static Icon getFolderIcon() {
@@ -61,22 +84,23 @@ public class FileUtility {
 	}
 	
 	public static String getType(String fileExt) {
-		if (cacheType.containsKey(fileExt))
-			return cacheType.get(fileExt);
-		
-		File f = new File(temp, "type." + fileExt);
-		try {
-			f.createNewFile();
-			f.deleteOnExit();
-			if (f.exists()) {
-				String i = fsv.getSystemTypeDescription(f);
-				cacheType.put(fileExt, i);
-				return i;
+		if (WINDOWS) {
+			if (cacheType.containsKey(fileExt))
+				return cacheType.get(fileExt);
+			
+			File f = new File(temp, "type." + fileExt);
+			try {
+				f.createNewFile();
+				f.deleteOnExit();
+				if (f.exists()) {
+					String i = fsv.getSystemTypeDescription(f);
+					cacheType.put(fileExt, i);
+					return i;
+				}
+			} catch (IOException e) {
+				_log.log(Level.WARNING, "Error creating temp file for type description", e);
 			}
-		} catch (IOException e) {
-			_log.log(Level.WARNING, "Error creating temp file for type description", e);
 		}
-		
 		return "";
 	}
 }
